Updated minor positiong of elements in diffrent resolution
